[bugfix] Don't attempt to download undownloadable files
Slack treats some links as files, such as when a Google Doc is
linked in a conversation. The downloader will try to download
such files but fail because they have no url_private_download
property.

This change fixes this bug by skipping all files that do not
have a url_private_download attribute.