Add support for fetching message replies
The script now supports fetching message replies *by default*. I
chose to make it by default because currently the documentation
does not mention the inability to fetch replies, so my
assumption is that users would expect this functionality.

One of the downsides to downloading replies is that for each
message that has replies, a separate Invoke-WebRequest must be
made to fetch the replies for that message. This significantly
increases the number of WebRequests that must be made, since
in a typical Slack workspace there are lots of messages with
replies, but each message typically has a small (<20) number
of replies.

I have added a -NoReplies flag to prevent replies from being
downloaded.